Fix self-destruct by outage, add /rebuild

+++ src/main/kotlin/xyz/redslime/releaseradar/listener/QuitListener.kt
@@ -3,6 +3,7 @@ package xyz.redslime.releaseradar.listener
 import dev.kord.core.Kord
 import dev.kord.core.event.guild.GuildDeleteEvent
 import dev.kord.core.on
+import xyz.redslime.releaseradar.DiscordClient
 import xyz.redslime.releaseradar.asLong
 import xyz.redslime.releaseradar.db
 
@@ -14,7 +15,11 @@ class QuitListener {
 
     fun register(client: Kord) {
         client.on<GuildDeleteEvent> {
-            db.purgeServerData(guildId.asLong())
+            if(this.unavailable) {
+                DiscordClient.disconnectedGuilds.add(this.guildId)
+            } else {
+                db.purgeServerData(guildId.asLong())
+            }
         }
     }
 }
